requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
attrs==25.3.0
bcrypt==4.3.0
beautifulsoup4==4.13.3
bidict==0.23.1
blis==1.3.0
breadability==0.1.20
catalogue==2.0.10
cattrs==24.1.3
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.0
colorama==0.4.6
comm==0.2.2
confection==0.1.5
contourpy==1.3.1
cssselect==1.3.0
curl_cffi==0.10.0
cycler==0.12.1
cymem==2.0.11
debugpy==1.8.14
decorator==5.2.1
dnspython==2.7.0
docopt==0.6.2
email_validator==2.2.0
en_core_web_lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.8.0/en_core_web_lg-3.8.0-py3-none-any.whl#sha256=293e9547a655b25499198ab15a525b05b9407a75f10255e405e8c3854329ab63
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
executing==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.18.0
fonttools==4.57.0
frozendict==2.4.6
fsspec==2025.3.2
fuzzywuzzy==0.18.0
goose3==3.1.19
h11==0.14.0
httpcore==1.0.8
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
ipykernel==6.29.5
ipython==9.1.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jieba3k==0.35.1
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.8
langcodes==3.5.0
langdetect==1.0.9
language_data==1.3.0
lxml==5.3.2
lxml_html_clean==0.4.2
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mongoengine==0.29.1
mpmath==1.3.0
multitasking==0.0.11
murmurhash==1.0.12
nest-asyncio==1.6.0
networkx==3.4.2
newspaper3k==0.2.8
nltk==3.9.1
numpy==2.2.4
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
parso==0.8.4
passlib==1.7.4
peewee==3.17.9
pillow==11.2.1
platformdirs==4.3.7
preshed==3.0.9
prompt_toolkit==3.0.51
protobuf==5.29.4
psutil==7.0.0
pure_eval==0.2.3
pyahocorasick==2.1.0
pycountry==24.6.1
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
PyJWT==2.10.1
pymongo==4.12.0
pyparsing==3.2.3
pyrate-limiter==2.10.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-engineio==4.12.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
RapidFuzz==3.13.0
regex==2024.11.6
requests==2.32.3
requests-cache==1.2.1
requests-file==2.1.0
requests-ratelimiter==0.7.0
rich==14.0.0
rich-toolkit==0.14.1
safetensors==0.5.3
schedule==1.2.2
scikit-learn==1.6.1
scipy==1.15.2
selenium==4.31.0
setuptools==78.1.0
sgmllib3k==1.0.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
spacy==3.8.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
stack-data==0.6.3
starlette==0.46.2
sumy==0.11.0
sympy==1.13.1
thinc==8.3.6
threadpoolctl==3.6.0
tinysegmenter==0.3
tldextract==5.2.0
tokenizers==0.21.1
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.3
trio==0.29.0
trio-websocket==0.12.2
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
url-normalize==2.2.1
urllib3==2.4.0
uvicorn==0.34.1
wasabi==1.1.3
watchfiles==1.0.5
wcwidth==0.2.13
weasel==0.4.1
websocket-client==1.8.0
websockets==15.0.1
wheel==0.45.1
wrapt==1.17.2
wsproto==1.2.0
yfinance==0.2.59
